refactor on invoice rejected
