+++ accounting-domain/src/test/java/onlydust/com/marketplace/accounting/domain/observer/AccountingNotifierTest.java
@@ -11,9 +11,11 @@ import onlydust.com.marketplace.accounting.domain.model.user.GithubUserId;
 import onlydust.com.marketplace.accounting.domain.model.user.UserId;
 import onlydust.com.marketplace.accounting.domain.port.out.AccountingRewardStoragePort;
 import onlydust.com.marketplace.accounting.domain.port.out.BillingProfileStoragePort;
+import onlydust.com.marketplace.accounting.domain.port.out.InvoiceStoragePort;
 import onlydust.com.marketplace.accounting.domain.service.AccountBookFacade;
 import onlydust.com.marketplace.accounting.domain.service.AccountingNotifier;
 import onlydust.com.marketplace.accounting.domain.view.*;
+import onlydust.com.marketplace.kernel.exception.OnlyDustException;
 import onlydust.com.marketplace.kernel.model.RewardStatus;
 import onlydust.com.marketplace.kernel.model.blockchain.evm.ethereum.Name;
 import onlydust.com.marketplace.kernel.model.blockchain.evm.ethereum.WalletLocator;
@@ -21,6 +23,7 @@ import onlydust.com.marketplace.kernel.port.output.NotificationPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
 
 import java.math.BigDecimal;
 import java.time.ZonedDateTime;
@@ -32,17 +35,18 @@ import static onlydust.com.marketplace.accounting.domain.AccountBookTest.account
 import static onlydust.com.marketplace.accounting.domain.stubs.BillingProfileHelper.newKyb;
 import static onlydust.com.marketplace.accounting.domain.stubs.Currencies.ETH;
 import static onlydust.com.marketplace.accounting.domain.stubs.Currencies.USD;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.*;
 
 public class AccountingNotifierTest {
     private BillingProfileStoragePort billingProfileStoragePort;
     AccountingNotifier accountingNotifier;
     NotificationPort mailObserver;
+    InvoiceStoragePort invoiceStoragePort;
     AccountingRewardStoragePort accountingRewardStoragePort;
     NotificationPort notificationPort;
     final Faker faker = new Faker();
-    final Currency currency = ETH;
-    final BigDecimal rewardAmount = BigDecimal.valueOf(faker.number().randomNumber(3, true));
 
     @BeforeEach
     void setUp() {
@@ -50,7 +54,8 @@ public class AccountingNotifierTest {
         mailObserver = mock(NotificationPort.class);
         accountingRewardStoragePort = mock(AccountingRewardStoragePort.class);
         notificationPort = mock(NotificationPort.class);
-        accountingNotifier = new AccountingNotifier(billingProfileStoragePort, accountingRewardStoragePort, mailObserver, notificationPort);
+        invoiceStoragePort = mock(InvoiceStoragePort.class);
+        accountingNotifier = new AccountingNotifier(billingProfileStoragePort, accountingRewardStoragePort, invoiceStoragePort, mailObserver, notificationPort);
     }
 
     @Nested
@@ -221,33 +226,54 @@ public class AccountingNotifierTest {
                     new Invoice.Reward(RewardId.random(), ZonedDateTime.now().minusDays(1), faker.lordOfTheRings().location(),
                             Money.of(BigDecimal.ONE, ETH), Money.of(2700L, USD), invoice.id(), List.of())
             ));
+
+            when(invoiceStoragePort.get(invoice.id())).thenReturn(Optional.of(invoice));
         }
 
         @Test
         public void should_update_reward_status_data() {
             // Given
-            final InvoiceRejected invalidInvoice = new InvoiceRejected(
-                    faker.internet().emailAddress(),
-                    3L,
-                    faker.internet().slug(),
-                    faker.name().firstName(),
-                    UUID.randomUUID(),
-                    invoice.number().toString(),
-                    invoice.rewards().stream().map(r -> ShortReward.builder()
-                            .id(r.id())
-                            .projectName(r.projectName())
-                            .currencyCode(currency.code().toString())
-                            .amount(r.amount().getValue())
-                            .dollarsEquivalent(r.amount().getValue())
-                            .build()).toList(),
-                    "Invalid invoice"
-            );
+            final var invoiceCreator = BillingProfileCoworkerView.builder()
+                    .userId(UserId.random())
+                    .login(faker.name().username())
+                    .email(faker.internet().emailAddress())
+                    .firstName(faker.name().firstName())
+                    .githubUserId(GithubUserId.of(faker.number().randomNumber(10, true)))
+                    .role(BillingProfile.User.Role.ADMIN)
+                    .joinedAt(ZonedDateTime.now())
+                    .invitedAt(null)
+                    .rewardCount(0)
+                    .billingProfileAdminCount(1)
+                    .build();
+
+            when(billingProfileStoragePort.findBillingProfileAdmin(invoice.createdBy(), invoice.billingProfileSnapshot().id()))
+                    .thenReturn(Optional.of(invoiceCreator));
 
             // When
-            accountingNotifier.onInvoiceRejected(invalidInvoice);
+            accountingNotifier.onInvoiceRejected(invoice.id(), "Invalid invoice");
 
             // Then
-            verify(mailObserver, times(1)).notify(invalidInvoice);
+            final var rejectedArgumentCaptor = ArgumentCaptor.forClass(InvoiceRejected.class);
+            verify(mailObserver).notify(rejectedArgumentCaptor.capture());
+            assertThat(rejectedArgumentCaptor.getValue().rejectionReason()).isEqualTo("Invalid invoice");
+            assertThat(rejectedArgumentCaptor.getValue().invoiceName()).isEqualTo(invoice.number().value());
+            assertThat(rejectedArgumentCaptor.getValue().billingProfileAdminEmail()).isEqualTo(invoiceCreator.email());
+            assertThat(rejectedArgumentCaptor.getValue().billingProfileAdminFirstName()).isEqualTo(invoiceCreator.firstName());
+            assertThat(rejectedArgumentCaptor.getValue().billingProfileAdminGithubLogin()).isEqualTo(invoiceCreator.login());
+            assertThat(rejectedArgumentCaptor.getValue().rewardCount()).isEqualTo(invoice.rewards().size());
+            assertThat(rejectedArgumentCaptor.getValue().rewards().get(0).getProjectName()).isEqualTo(invoice.rewards().get(0).projectName());
+            assertThat(rejectedArgumentCaptor.getValue().rewards().get(0).getAmount()).isEqualTo(invoice.rewards().get(0).amount().getValue());
+            assertThat(rejectedArgumentCaptor.getValue().rewards().get(0).getCurrencyCode()).isEqualTo(invoice.rewards().get(0).amount().getCurrency().code().toString());
+
+        }
+
+        @Test
+        void should_not_send_notification_given_a_billing_profile_admin_not_found() {
+            // When
+            assertThatThrownBy(() -> accountingNotifier.onInvoiceRejected(invoice.id(), "Invalid invoice"))
+                    // Then
+                    .isInstanceOf(OnlyDustException.class)
+                    .hasMessage("Billing profile admin not found for billing profile %s".formatted(invoice.billingProfileSnapshot().id()));
         }
     }
